refactor: change components folder naming
