+++ src/components/form/_index.ts
@@ -1,2 +1,2 @@
-export { Form } from '~compos/form/form/Form';
-export { UsersList } from '~compos/form/users/usersList/UsersList';
+export { Form } from '@/components/form/Form/Form';
+export { UsersList } from '@/components/form/users/UsersList/UsersList';
