+++ src/components/form/fields/_index.ts
@@ -1,6 +1,6 @@
-export { AddInfo } from '~compos/form/fields/addInfo/AddInfo';
-export { BirthDate } from '~compos/form/fields/birthDate/BirthDate';
-export { Country } from '~compos/form/fields/country/Country';
-export { Gender } from '~compos/form/fields/gender/Gender';
-export { Photo } from '~compos/form/fields/photo/Photo';
-export { UserName } from '~compos/form/fields/userName/userName';
+export { AddInfo } from '@/components/form/fields/AddInfo/AddInfo';
+export { BirthDate } from '@/components/form/fields/BirthDate/BirthDate';
+export { Country } from '@/components/form/fields/Country/Country';
+export { Gender } from '@/components/form/fields/Gender/Gender';
+export { Photo } from '@/components/form/fields/Photo/Photo';
+export { UserName } from '@/components/form/fields/UserName/userName';
