+++ src/components/form/fields/birthDate/BirthDate.tsx
@@ -1,6 +1,6 @@
 import { type FC } from 'react';
 
-import { FormError } from '../error/FormError';
+import { FormError } from '../Error/FormError';
 
 import styles from '~global/scss/Input.module.scss';
 import '../styles/Field.scss';
