+++ src/components/form/fields/userName/userName.tsx
@@ -1,5 +1,5 @@
 import { type FC } from 'react';
-import { FormError } from '../error/FormError';
+import { FormError } from '../Error/FormError';
 
 import styles from '~global/scss/Input.module.scss';
 import '../styles/Field.scss';
