+++ src/services/storage/localStorage.ts
@@ -1,4 +1,4 @@
-import { StorageService } from '@/services/storage';
+import { StorageService } from '@/services/storage/storage';
 
 export class LocalStorageService extends StorageService {
   public constructor(prefix: string) {
